+++ tools/cli/src/options.hpp
@@ -27,6 +27,8 @@ public:
     int         color = 0;
     int         hex_input = 0;
     int         pdol = 0;
+    int         decode = 0;
+    int         json = 0;
 };
 
 } // namespace cli